Moves wait_option type & support out to its own files
